Add button to DrawingComponent to enable selectAndModify tool in MapDrawingsButton using @ViewChild, integrate CreateDrawingButtonComponent into DrawingComponent to use direct view child
