--- projects/core/src/lib/components/drawing/create-drawing-button/create-drawing-button.component.css
@@ -1,3 +1,0 @@
-p {
-  margin-bottom: 0;
-}
