+++ projects/core/src/lib/map/components/map-drawing-buttons/map-drawing-buttons.component.ts
@@ -56,6 +56,9 @@ export class MapDrawingButtonsComponent implements OnInit, OnDestroy {
   public allowModify = false;
 
   @Input()
+  public showSelectButton = true;
+
+  @Input()
   public allowedShapes: DrawingFeatureTypeEnum[] | undefined = undefined;
 
   @Output()
